+++ QW/client/cl_tent.c
@@ -385,7 +385,7 @@ CL_UpdateBeams(void)
 	    if (yaw < 0)
 		yaw += 360;
 
-	    forward = sqrt(dist[0] * dist[0] + dist[1] * dist[1]);
+	    forward = sqrtf(dist[0] * dist[0] + dist[1] * dist[1]);
 	    pitch = (int)(atan2(dist[2], forward) * 180 / M_PI);
 	    if (pitch < 0)
 		pitch += 360;
