+++ common/d_polyse.c
@@ -526,9 +526,9 @@ D_PolysetCalcGradients(int skinwidth)
     t0 = r_p0[4] - r_p2[4];
     t1 = r_p1[4] - r_p2[4];
     r_lstepx = (int)
-	ceil((t1 * p01_minus_p21 - t0 * p11_minus_p21) * xstepdenominv);
+	ceilf((t1 * p01_minus_p21 - t0 * p11_minus_p21) * xstepdenominv);
     r_lstepy = (int)
-	ceil((t1 * p00_minus_p20 - t0 * p10_minus_p20) * ystepdenominv);
+	ceilf((t1 * p00_minus_p20 - t0 * p10_minus_p20) * ystepdenominv);
 
     t0 = r_p0[2] - r_p2[2];
     t1 = r_p1[2] - r_p2[2];
