+++ common/d_sprite.c
@@ -209,13 +209,13 @@ D_SpriteScanLeftEdge(void)
     if (lmaxindex == 0)
 	lmaxindex = r_spritedesc.nump;
 
-    vtop = ceil(r_spritedesc.pverts[i].v);
+    vtop = ceilf(r_spritedesc.pverts[i].v);
 
     do {
 	pvert = &r_spritedesc.pverts[i];
 	pnext = pvert - 1;
 
-	vbottom = ceil(pnext->v);
+	vbottom = ceilf(pnext->v);
 
 	if (vtop < vbottom) {
 	    du = pnext->u - pvert->u;
@@ -269,7 +269,7 @@ D_SpriteScanRightEdge(void)
     if (vvert > r_refdef.fvrectbottom_adj)
 	vvert = r_refdef.fvrectbottom_adj;
 
-    vtop = ceil(vvert);
+    vtop = ceilf(vvert);
 
     do {
 	pvert = &r_spritedesc.pverts[i];
@@ -281,7 +281,7 @@ D_SpriteScanRightEdge(void)
 	if (vnext > r_refdef.fvrectbottom_adj)
 	    vnext = r_refdef.fvrectbottom_adj;
 
-	vbottom = ceil(vnext);
+	vbottom = ceilf(vnext);
 
 	if (vtop < vbottom) {
 	    uvert = pvert->u;
@@ -408,8 +408,8 @@ D_DrawSprite(void)
 	pverts++;
     }
 
-    ymin = ceil(ymin);
-    ymax = ceil(ymax);
+    ymin = ceilf(ymin);
+    ymax = ceilf(ymax);
 
     if (ymin >= ymax)
 	return;			// doesn't cross any scans at all
