+++ common/r_surf.c
@@ -86,7 +86,7 @@ R_AddDynamicLights(void)
 	rad = cl_dlights[lnum].radius;
 	dist = DotProduct(cl_dlights[lnum].origin, surf->plane->normal) -
 	    surf->plane->dist;
-	rad -= fabs(dist);
+	rad -= fabsf(dist);
 	minlight = cl_dlights[lnum].minlight;
 	if (rad < minlight)
 	    continue;
